検索履歴をUserSetting.xmlに覚えて復元するように対応。
